+++ datacrunch/instances/instances.py
@@ -1,5 +1,6 @@
 from typing import List, Union, Optional, Dict
 from datacrunch.helpers import stringify_class_object_properties
+from datacrunch.constants import Locations
 
 INSTANCES_ENDPOINT = '/instances'
 
@@ -24,7 +25,7 @@ class Instance:
                  storage: dict,
                  os_volume_id: str,
                  gpu_memory: dict,
-                 location: str = "FIN1",
+                 location: str = Locations.FIN_01,
                  startup_script_id: str = None,
                  is_spot: bool = False
                  ) -> None:
@@ -62,7 +63,7 @@ class Instance:
         :type id: str
         :param memory: gpu memory details
         :type memory: dict
-        :param location: datacenter location, defaults to "FIN1"
+        :param location: datacenter location, defaults to "FIN-01"
         :type location: str, optional
         :param startup_script_id: startup script id, defaults to None
         :type startup_script_id: str, optional
@@ -268,6 +269,7 @@ class Instance:
         """
         return stringify_class_object_properties(self)
 
+
 class InstancesService:
     """A service for interacting with the instances endpoint"""
 
@@ -295,7 +297,8 @@ class InstancesService:
             ip=instance_dict['ip'],
             status=instance_dict['status'],
             created_at=instance_dict['created_at'],
-            ssh_key_ids=instance_dict['ssh_key_ids'] if 'ssh_key_ids' in instance_dict else [],
+            ssh_key_ids=instance_dict['ssh_key_ids'] if 'ssh_key_ids' in instance_dict else [
+            ],
             startup_script_id=instance_dict['startup_script_id'] if 'startup_script_id' in instance_dict else None,
             cpu=instance_dict['cpu'],
             gpu=instance_dict['gpu'],
@@ -328,7 +331,8 @@ class InstancesService:
             ip=instance_dict['ip'],
             status=instance_dict['status'],
             created_at=instance_dict['created_at'],
-            ssh_key_ids=instance_dict['ssh_key_ids'] if 'ssh_key_ids' in instance_dict else [],
+            ssh_key_ids=instance_dict['ssh_key_ids'] if 'ssh_key_ids' in instance_dict else [
+            ],
             startup_script_id=instance_dict['startup_script_id'] if 'startup_script_id' in instance_dict else None,
             cpu=instance_dict['cpu'],
             gpu=instance_dict['gpu'],
@@ -346,7 +350,7 @@ class InstancesService:
                hostname: str,
                description: str,
                ssh_key_ids: list = [],
-               location: str = "FIN1",
+               location: str = Locations.FIN_01,
                startup_script_id: str = None,
                volumes: List[Dict] = None,
                existing_volumes: List[str] = None,
@@ -365,7 +369,7 @@ class InstancesService:
         :type hostname: str
         :param description: instance description
         :type description: str
-        :param location: datacenter location, defaults to "FIN1"
+        :param location: datacenter location, defaults to "FIN-01"
         :type location: str, optional
         :param startup_script_id: startup script id, defaults to None
         :type startup_script_id: str, optional
@@ -389,7 +393,7 @@ class InstancesService:
             "startup_script_id": startup_script_id,
             "hostname": hostname,
             "description": description,
-            "location": location,
+            "location_code": location,
             "os_volume": os_volume,
             "volumes": volumes,
             "existing_volumes": existing_volumes,