+++ tests/integration_tests/test_volumes.py
@@ -1,8 +1,12 @@
 import os
 import pytest
 from datacrunch.datacrunch import DataCrunchClient
+from datacrunch.constants import Locations, VolumeTypes, VolumeStatus
 
 IN_GITHUB_ACTIONS = os.getenv("GITHUB_ACTIONS") == "true"
+
+
+NVMe = VolumeTypes.NVMe
 
 
 @pytest.mark.skipif(IN_GITHUB_ACTIONS, reason="Test doesn't work in Github Actions.")
@@ -12,7 +16,7 @@ class TestVolumes():
     def test_get_volumes_from_trash(self, datacrunch_client: DataCrunchClient):
         # create new volume
         volume = datacrunch_client.volumes.create(
-            type=datacrunch_client.constants.volume_types.NVMe, name="test_volume", size=100)
+            type=NVMe, name="test_volume", size=100)
 
         # delete volume
         datacrunch_client.volumes.delete(volume.id)
@@ -29,7 +33,7 @@ class TestVolumes():
     def test_permanently_delete_detached_volumes(seld, datacrunch_client):
         # create new volume
         volume = datacrunch_client.volumes.create(
-            type=datacrunch_client.constants.volume_types.NVMe, name="test_volume", size=100)
+            type=NVMe, name="test_volume", size=100)
 
         # permanently delete the detached volume
         datacrunch_client.volumes.delete(volume.id, is_permanent=True)
@@ -49,7 +53,7 @@ class TestVolumes():
     def test_permanently_delete_a_deleted_volume_from_trash(self, datacrunch_client):
         # create new volume
         volume = datacrunch_client.volumes.create(
-            type=datacrunch_client.constants.volume_types.NVMe, name="test_volume", size=100)
+            type=NVMe, name="test_volume", size=100)
 
         # delete volume
         datacrunch_client.volumes.delete(volume.id)
@@ -68,3 +72,16 @@ class TestVolumes():
 
         # assert volume is not in trash
         assert volume.id not in [v.id for v in volumes]
+
+    def test_create_volume(self, datacrunch_client):
+        # create new volume
+        volume = datacrunch_client.volumes.create(
+            type=NVMe, name="test_volume", size=100, location=Locations.FIN_01)
+
+        # assert volume is created
+        assert volume.id is not None
+        assert volume.location == Locations.FIN_01
+        assert volume.status == VolumeStatus.ORDERED or volume.status == VolumeStatus.DETACHED
+
+        # cleaning: delete volume
+        datacrunch_client.volumes.delete(volume.id, is_permanent=True)